+++ VDA-Reward-Contract/ChangeLog.md
@@ -1,3 +1,7 @@
+2024-01-30 (v0.5.0)
+-------------------
+- Added `getTokenAddress()` and `getStorageNodeContractAddress()` functions
+
 2024-01-23 (v0.4.1)
 -------------------
 - Fixed the deployment script